Fixed scope issue with version macro.
Due to Julia's importing mechanisms, it's impossible to use a builtin
like include() or require() to import code (and eval it) within the
context of a module that isn't Main.

The only path I can find around this problem is to make the same calls
that the require() function makes, but do it ourselves within the OpenGL
module context.

This fixes the issue with the `OpenGLStd` module ending up in the Main
namespace, instead of the OpenGL namespace. However, any attempt to
export these symbols throws them into the Main namespace. The current
implementation omits the export call, ensuring that this:

> using OpenGL
> @OpenGL.version "3.2"

Only ever populates the OpenGL module's namespace. It unfortunately also
prevents the use of a `using OpenGL` or `importall OpenGL` call to bring
the symbols into another namespace.

Using a custom macro for the import would resolve this, but it does mean
our OpenGL module never appears like a full, regular module, as it
doesn't export any symbols itself.

Alternatively, determining how to alter the OpenGL module so that its
exports include those of the modules we dynamically add to it would
allow for more natural behaviour.